Rework CLI into a plugin system (#29)
