+++ index.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import { runCLI } from './src/CLI';
+import { runCLI } from './src/CLI/CLI';
 
 void runCLI();
